Moved the file AuthorizeOperationMessageProcessor.cs to /Extensibility and changed its namespace to SoapCore.Extensibility
